+++ docs/_includes/CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.6
 
 * Added installed and upgraded pages
+* Added toolbar coloring feature